% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{approaches}
\alias{approaches}
\title{ICD-10-PCS Approaches}
\usage{
approaches(x = NULL)
}
\arguments{
\item{x}{ICD-10-PCS section character, an alphanumeric code of length 1.
If \code{NULL} (default), returns all 17 sections.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
ICD-10-PCS Approaches
}
\examples{

approaches()

approaches("0")

approaches("X")

}
