% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{code_range}
\alias{code_range}
\title{Return a range of ICD-10-PCS codes}
\usage{
code_range(start, end)
}
\arguments{
\item{start}{7-character ICD-10-PCS code.}

\item{end}{7-character ICD-10-PCS code.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Return a range of ICD-10-PCS codes
}
\examples{
code_range("0G9000Z", "0G9100Z")

code_range("0G9000Z", "10D20ZZ")

}
