% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapser}
\alias{collapser}
\title{Wrapper for \code{\link[=paste0]{paste0()}} that collapses result}
\usage{
collapser(x)
}
\arguments{
\item{x}{split character vector}
}
\value{
collapsed character vector
}
\description{
Wrapper for \code{\link[=paste0]{paste0()}} that collapses result
}
\keyword{internal}
