% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions}
\alias{definitions}
\title{ICD-10-PCS Definitions}
\usage{
definitions()
}
\value{
\code{\link[=tibble]{tibble()}}
}
\description{
ICD-10-PCS Definitions
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
definitions()
\dontshow{\}) # examplesIf}
}
