% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delister}
\alias{delister}
\title{Wrapper for \code{\link[=unlist]{unlist()}}, with \code{use.names} set to \code{FALSE}}
\usage{
delister(x)
}
\arguments{
\item{x}{character vector}
}
\value{
split character vector
}
\description{
Wrapper for \code{\link[=unlist]{unlist()}}, with \code{use.names} set to \code{FALSE}
}
\keyword{internal}
