% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{devices}
\alias{devices}
\title{ICD-10-PCS Index}
\usage{
devices(system = NULL, device = NULL)
}
\arguments{
\item{system}{Column to search}

\item{device}{description}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
ICD-10-PCS Index
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
devices()
\dontshow{\}) # examplesIf}
}
