% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_print.R
\name{gt_pcs2}
\alias{gt_pcs2}
\title{Return ICD-10-PCS Code as a \code{gt} Table}
\usage{
gt_pcs2(x)
}
\arguments{
\item{x}{an object of class \code{pcs}.}
}
\value{
description
}
\description{
Return ICD-10-PCS Code as a \code{gt} Table
}
\examples{
pcs("0G9000Z") |> gt_pcs2()

}
