% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{includes}
\alias{includes}
\title{ICD-10-PCS Includes}
\usage{
includes(section = NULL, axis = NULL, name = NULL, text = NULL)
}
\arguments{
\item{section}{PCS section character.}

\item{axis}{PCS axis position.}

\item{name}{PCS axis name.}

\item{text}{PCS axis label.}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
ICD-10-PCS Includes
}
\examples{
includes(section = "0", axis = "3")
}
