% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{index}
\alias{index}
\title{ICD-10-PCS Index}
\usage{
index(search = NULL, column = NULL)
}
\arguments{
\item{search}{Search term}

\item{column}{Column to search}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
ICD-10-PCS Index
}
\examples{
index(search = "Abdominohysterectomy")

index(search = "Attain Ability")

index(search = "radi")

}
