% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is_section_0}
\alias{is_section_0}
\title{Validate Section 0 Code}
\usage{
is_section_0(pcs_code)
}
\arguments{
\item{pcs_code}{\verb{<chr>} string}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, otherwise \code{FALSE}
}
\description{
Section 0 (Medical and Surgical) accounts for approx. 87\% of all PCS codes.
}
\details{
\code{68058 / 78603} = \code{0.8658448}
\subsection{Valid Code Structure}{

A valid Section 0 Code is 7 characters long and follows the pattern:
\enumerate{
\item \verb{[0]}
\item \verb{[B-DF-HJ-NP-Y0-9]}
\item \verb{[B-DF-HJ-NP-Y0-9]}
\item \verb{[A-HJ-NP-Y0-9]}
\item \verb{[FX03478]}
\item \verb{[A-HJ-NP-Z0-9]}
\item \verb{[A-HJ-NP-Z0-9]}
}
}
}
\examples{
x <- c("00DN0ZZ", "01UB3KZ", "0BBBFBB")

is_section_0(x)

x[which(is_section_0(x))]

try(is_section_0("1164"))

}
