% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_codes.R
\name{is_section_B}
\alias{is_section_B}
\title{Validate Imaging Codes}
\usage{
is_section_B(pcs_code)
}
\arguments{
\item{pcs_code}{\verb{<chr>} string}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, otherwise \code{FALSE}
}
\description{
Section B accounts for 3.79\% of all PCS codes.
}
\examples{
x <- c("00DN0ZZ", "01UB3KZ", "0BBBFBB")

is_section_B(x)

x[which(is_section_B(x))]

try(is_section_B("01UB3K"))

}
\seealso{
Other PCS code validators: 
\code{\link{is_section_0}()},
\code{\link{is_section_1}()},
\code{\link{is_section_2}()},
\code{\link{is_section_3}()},
\code{\link{is_section_4}()},
\code{\link{is_section_5}()},
\code{\link{is_section_6}()},
\code{\link{is_section_7}()},
\code{\link{is_section_8}()},
\code{\link{is_section_9}()},
\code{\link{is_section_C}()},
\code{\link{is_section_D}()},
\code{\link{is_section_F}()},
\code{\link{is_section_G}()},
\code{\link{is_section_H}()},
\code{\link{is_section_X}()},
\code{\link{is_section_code}()}
}
\concept{PCS code validators}
