% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_codes.R
\name{is_section_code}
\alias{is_section_code}
\title{Validate Full PCS Codes by Section}
\usage{
is_section_code(section_value = substr(pcs_code, 1, 1)[1], pcs_code)
}
\arguments{
\item{section_value}{\verb{<chr>} string}

\item{pcs_code}{\verb{<chr>} string}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, otherwise \code{FALSE}
}
\description{
Validate Full PCS Codes by Section
}
\examples{
is_section_code(section_value = "0", pcs_code = "01UB3KZ")

x <- c("00DN0ZZ", "01UB3KZ", "0BBBFBB")

is_section_code(pcs_code = x)

x[which(is_section_code(pcs_code = x))]

try(is_section_code(pcs_code = "01UB3K"))

}
\concept{PCS code validators}
