% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_values.R
\name{is_section_value}
\alias{is_section_value}
\title{Validate Section Values}
\usage{
is_section_value(pcs_value)
}
\arguments{
\item{pcs_value}{\verb{<chr>} string}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, otherwise \code{FALSE}
}
\description{
The 17 valid PCS Section values are "B-D", "F-H", "X", and "0-9".
}
\examples{
x <- c("0", "3", "K")

is_section_value(x)

x[which(is_section_value(x))]

try(is_section_value("XX"))

}
\seealso{
Other Value Validators: 
\code{\link{is_system_value}()}
}
\concept{Value Validators}
