% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_codes.R
\name{is_section_X}
\alias{is_section_X}
\title{Validate New Technology Codes}
\usage{
is_section_X(pcs_code)
}
\arguments{
\item{pcs_code}{\verb{<chr>} string}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, otherwise \code{FALSE}
}
\description{
Section X accounts for 0.44\% of all PCS codes.
}
\details{
\subsection{Code Pattern}{

Section X codes adhere to the following the pattern:
\enumerate{
\item \verb{[1]}
\item \verb{[0]}
\item \verb{[ADEHJPQSTY29]}
\item \verb{[0-2]}
\item \verb{[X03478]}
\item \verb{[YZ3]}
\item \verb{[A-HJ-NP-Z0-9]}
}
}
}
\examples{
x <- c("00DN0ZZ", "01UB3KZ", "0BBBFBB")

is_section_X(x)

x[which(is_section_X(x))]

try(is_section_X("01UB3K"))

}
\seealso{
Other Section Validators: 
\code{\link{is_section_0}()},
\code{\link{is_section_1}()},
\code{\link{is_section_2}()},
\code{\link{is_section_3}()},
\code{\link{is_section_4}()},
\code{\link{is_section_5}()},
\code{\link{is_section_6}()},
\code{\link{is_section_7}()},
\code{\link{is_section_8}()},
\code{\link{is_section_9}()},
\code{\link{is_section_B}()},
\code{\link{is_section_C}()},
\code{\link{is_section_D}()},
\code{\link{is_section_F}()},
\code{\link{is_section_G}()},
\code{\link{is_section_H}()}
}
\concept{Section Validators}
