% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{operations}
\alias{operations}
\title{ICD-10-PCS Root Operations}
\usage{
operations()
}
\value{
\code{\link[=tibble]{tibble()}}
}
\description{
ICD-10-PCS Root Operations
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
operations()
\dontshow{\}) # examplesIf}
}
