% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs.R
\name{pcs}
\alias{pcs}
\title{Look up ICD-10-PCS Codes}
\usage{
pcs(x)
}
\arguments{
\item{x}{ICD-10-PCS code, a \strong{7-character} alphanumeric code}
}
\value{
\code{\link[=tibble]{tibble()}}
}
\description{
Look up ICD-10-PCS Codes
}
\examples{
pcs("0G9000Z")

pcs("0016070")

pcs("2W0UX0Z")

pcs("2W20X4Z")

pcs("2W1HX7Z")

}
