% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs_build_code.R
\name{pcs_code}
\alias{pcs_code}
\title{Look up ICD-10-PCS Codes or Tables}
\usage{
pcs_code(x = NULL)
}
\arguments{
\item{x}{an alphanumeric character vector, can be 3 to 7 characters long.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Look up ICD-10-PCS Codes or Tables
}
\examples{

pcs_code("0G9")

pcs_code("0G90")

pcs_code("0G900")

pcs_code("0G9000")

pcs_code("0G9000Z")

}
