% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rows}
\alias{rows}
\title{ICD-10-PCS Rows}
\usage{
rows(x = NULL)
}
\arguments{
\item{x}{3 to 4-character string representing an ICD-10-PCS row within a table.
If \code{NULL} (default), returns all ~ 29k rows.}
}
\value{
\code{\link[=tibble]{tibble()}}
}
\description{
ICD-10-PCS Rows
}
\examples{
rows("00X")

rows("00XF")

}
