% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{section}
\alias{section}
\title{Look up ICD-10-PCS Section}
\usage{
section(x = NULL)
}
\arguments{
\item{x}{ICD-10-PCS section character, an alphanumeric code of length 1.
If \code{NULL} (default), returns all 17 sections.}
}
\value{
\code{\link[=tibble]{tibble()}}
}
\description{
Look up ICD-10-PCS Section
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
section()
section("0")
section("2")
section("X")
\dontshow{\}) # examplesIf}
}
