% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{sections}
\alias{sections}
\title{ICD-10-PCS Sections}
\usage{
sections(x = NULL)
}
\arguments{
\item{x}{ICD-10-PCS section character, an alphanumeric code of length 1.
If \code{NULL} (default), returns all 17 sections.}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
ICD-10-PCS Sections
}
\examples{
sections()

sections("0")

sections("2")

sections("X")

}
