% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitter}
\alias{splitter}
\title{Wrapper for \code{\link[=strsplit]{strsplit()}} that unlists result}
\usage{
splitter(x)
}
\arguments{
\item{x}{character vector}
}
\value{
split character vector
}
\description{
Wrapper for \code{\link[=strsplit]{strsplit()}} that unlists result
}
\keyword{internal}
