% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{srchcol}
\alias{srchcol}
\title{Search a data frame's column by string}
\usage{
srchcol(df, col, search, ignore = FALSE)
}
\arguments{
\item{df}{data frame}

\item{col}{column to search}

\item{search}{string to search}

\item{ignore}{ignore case or not}
}
\value{
collapsed character vector
}
\description{
Search a data frame's column by string
}
\keyword{internal}
