% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systems.R
\name{systems}
\alias{systems}
\title{ICD-10-PCS Body Systems}
\usage{
systems()
}
\value{
\code{\link[=tibble]{tibble()}}
}
\description{
ICD-10-PCS Body Systems
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
systems()
\dontshow{\}) # examplesIf}
}
