% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{table}
\alias{table}
\title{ICD-10-PCS Tables}
\usage{
table(x = NULL)
}
\arguments{
\item{x}{1 to 3-character string representing an ICD-10-PCS table.
If \code{NULL} (default), returns all 895 tables.}
}
\value{
\code{\link[=tibble]{tibble()}}
}
\description{
ICD-10-PCS Tables
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
table()
table("0")
table("00")
table("00X")
\dontshow{\}) # examplesIf}
}
