% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tables}
\alias{tables}
\title{ICD-10-PCS Tables}
\usage{
tables(code = NULL, text = NULL)
}
\arguments{
\item{code}{1 to 3-character string representing an ICD-10-PCS table.
If \code{NULL} (default), returns all 895 tables.}

\item{text}{Search code descriptions}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
ICD-10-PCS Tables
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tables(code = "0")

tables(code = "00")

tables(code = "00X")
\dontshow{\}) # examplesIf}
}
