% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach.R
\name{approaches}
\alias{approaches}
\title{ICD-10-PCS Sections}
\usage{
approaches(x = NULL)
}
\arguments{
\item{x}{ICD-10-PCS section character, an alphanumeric code of length 1.
If \code{NULL} (default), returns all 17 sections.}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
ICD-10-PCS Sections
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

approaches()

approaches("0")

approaches("2")

approaches("X")
\dontshow{\}) # examplesIf}
}
