% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions}
\alias{definitions}
\title{ICD-10-PCS Definitions}
\usage{
definitions(section = NULL)
}
\arguments{
\item{section}{PCS section character.}
}
\value{
\code{\link[=tibble]{tibble()}}
}
\description{
ICD-10-PCS Definitions
}
\examples{
definitions(section = "2")

definitions(section = "4")

}
