% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions}
\alias{definitions}
\title{ICD-10-PCS Definitions}
\usage{
definitions(section = NULL, axis = NULL, value = NULL)
}
\arguments{
\item{section}{PCS section character.}

\item{axis}{PCS axis position.}

\item{value}{PCS axis value.}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
ICD-10-PCS Definitions
}
\examples{
definitions(section = "0", axis = "3")

definitions(section = "0", axis = "3", value = "Y")

definitions(section = "0", axis = "3", value = "Y")$elements

}
