% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions}
\alias{definitions}
\title{ICD-10-PCS Definitions}
\usage{
definitions(
  section = NULL,
  axis = NULL,
  col = c("value", "label", "name", "definition", "explanation"),
  search = NULL
)
}
\arguments{
\item{section}{PCS section character.}

\item{axis}{PCS axis position.}

\item{col}{column to search: "name", "value" (default), "label" "definition", "explanation"}

\item{search}{string to search for in \code{col}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
ICD-10-PCS Definitions
}
\examples{
definitions(section = "0", axis = "3", search = "Drainage")

definitions(section = "0", axis = "5")

definitions(section = "B", search = "Fluoroscopy")

}
