% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions}
\alias{definitions}
\title{ICD-10-PCS Definitions}
\usage{
definitions(section = NULL, axis = NULL, label = NULL)
}
\arguments{
\item{section}{PCS section character.}

\item{axis}{PCS axis position.}

\item{label}{PCS axis label.}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
ICD-10-PCS Definitions
}
\examples{
definitions(section = "0",
            axis = "3",
            label = "Drainage")

definitions(section = "0",
            axis = "3",
            label = "Drainage")$elements

definitions(section = "0",
            axis = "4",
            label = "Abdominal Aorta")$elements

definitions(section = "B",
            label = "Fluoroscopy")$elements

}
