% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{devices}
\alias{devices}
\title{ICD-10-PCS Devices}
\usage{
devices(
  system = NULL,
  operation = NULL,
  device = NULL,
  col = "device_name",
  search = NULL
)
}
\arguments{
\item{system}{PCS system character.}

\item{operation}{PCS operation character.}

\item{device}{PCS device character.}

\item{col}{column to search: "device_name" (default), "includes"}

\item{search}{string to search for in \code{col}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
ICD-10-PCS Devices
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
devices()
\dontshow{\}) # examplesIf}
}
