% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{includes}
\alias{includes}
\title{ICD-10-PCS Includes}
\usage{
includes(
  section = NULL,
  axis = NULL,
  col = c("label", "name", "includes"),
  search = NULL
)
}
\arguments{
\item{section}{PCS section:
\itemize{
\item \code{"0"} (Medical and Surgical),
\item \code{"3"} (Administration)
\item \code{"F"} (Physical Rehabilitation and Diagnostic Audiology)
\item \code{"G"} (Mental Health)
\item \code{"X"} (New Technology)
}}

\item{axis}{PCS axis:
\itemize{
\item \code{"3"}
\itemize{
\item Operation (0:9, X)
\item Type (B, C, F, G, H)
}
\item \code{"4"}
\itemize{
\item Qualifier (G)
}
\item \code{"5"}
\itemize{
\item Approach (0:4, 7:9, X)
\item Type Qualifier (F)
}
}}

\item{col}{column to search: \code{"name"}, \code{"label"} (default), \code{"includes"}}

\item{search}{string to search for in \code{col}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
ICD-10-PCS Includes
}
\examples{
includes(section = "0", axis = "3")

}
