% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{index}
\alias{index}
\title{ICD-10-PCS Index}
\usage{
index(search = NULL, col = "term")
}
\arguments{
\item{search}{string to search for in \code{col}}

\item{col}{column to search: "term" (default), "verb", "value", "code"}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
ICD-10-PCS Index
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
index(search = "Abdominohysterectomy")

index(search = "Attain Ability")

index(search = "radi*")
\dontshow{\}) # examplesIf}
}
