% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_codes.R
\name{is_section_code}
\alias{is_section_code}
\title{Validate Full PCS Codes by Section}
\usage{
is_section_code(section_value = substr(pcs_code, 1, 1)[1], pcs_code)
}
\arguments{
\item{section_value}{\verb{<chr>} string}

\item{pcs_code}{\verb{<chr>} string}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, otherwise \code{FALSE}
}
\description{
Validate Full PCS Codes by Section
}
\examples{
is_section_code(section_value = "0", pcs_code = "01UB3KZ")

x <- c("00DN0ZZ", "01UB3KZ", "0BBBFBB")

is_section_code(pcs_code = x)

x[which(is_section_code(pcs_code = x))]

try(is_section_code(pcs_code = "01UB3K"))

}
\seealso{
Other PCS code validators: 
\code{\link{is_section_0}()},
\code{\link{is_section_1}()},
\code{\link{is_section_2}()},
\code{\link{is_section_3}()},
\code{\link{is_section_4}()},
\code{\link{is_section_5}()},
\code{\link{is_section_6}()},
\code{\link{is_section_7}()},
\code{\link{is_section_8}()},
\code{\link{is_section_9}()},
\code{\link{is_section_B}()},
\code{\link{is_section_C}()},
\code{\link{is_section_D}()},
\code{\link{is_section_F}()},
\code{\link{is_section_G}()},
\code{\link{is_section_H}()},
\code{\link{is_section_X}()}
}
\concept{PCS code validators}
