% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_values.R
\name{is_system_value}
\alias{is_system_value}
\title{Validate System Values}
\usage{
is_system_value(pcs_value)
}
\arguments{
\item{pcs_value}{\verb{<chr>} string}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, otherwise \code{FALSE}
}
\description{
The 34 valid PCS System values are "A-H", "J-N", "P-Z", and "0-9".
}
\examples{
x <- c(LETTERS, 0:9)

is_system_value(x)

x[which(is_system_value(x))]

try(is_system_value("XX"))

}
\seealso{
Other Value Validators: 
\code{\link{is_section_value}()}
}
\concept{Value Validators}
