% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{order}
\alias{order}
\title{ICD-10-PCS Order File}
\usage{
order(col = "code", search = NULL)
}
\arguments{
\item{col}{column to search: "code" (default), "table", "row", "description_code", "description_table"}

\item{search}{string to search for in \code{col}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
ICD-10-PCS Order File
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
order(search = "00X")

order(search = "Olfactory", col = "description_code")
\dontshow{\}) # examplesIf}
}
