% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{order}
\alias{order}
\title{ICD-10-PCS Code Order}
\usage{
order(code = NULL, text = NULL)
}
\arguments{
\item{code}{1 to 7-character string.
If \code{NULL} (default), returns all 78,603 codes.}

\item{text}{Search code descriptions}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
ICD-10-PCS Code Order
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
order(code = "00X")

order(text = "Olfactory")
\dontshow{\}) # examplesIf}
}
