% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs.R
\name{pcs}
\alias{pcs}
\title{Look up ICD-10-PCS Codes or Tables}
\usage{
pcs(x)
}
\arguments{
\item{x}{an alphanumeric character vector, can be 3 to 7 characters long.}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Look up ICD-10-PCS Codes or Tables
}
\examples{
pcs("0G9")

pcs("0G90")

pcs("0G900")

pcs("0G9000")

pcs("0G9000Z")

}
