% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs.R
\name{pcs}
\alias{pcs}
\title{Look up ICD-10-PCS Codes}
\usage{
pcs(x)
}
\arguments{
\item{x}{ICD-10-PCS code, a \strong{7-character} alphanumeric code}
}
\value{
description
}
\description{
Look up ICD-10-PCS Codes
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pcs("2W0UX0Z")
pcs("2W20X4Z")
pcs("2W10X7Z")
pcs("2W10X6Z")
pcs("2W1HX7Z")
pcs("0G9000Z")
pcs("0016070")
\dontshow{\}) # examplesIf}
}
