% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedural-package.R
\docType{package}
\name{procedural-package}
\alias{procedural}
\alias{procedural-package}
\title{procedural: Tidy ICD-10-PCS interface}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a tidy interface to the ICD-10-PCS procedure code system. The package includes a dataset of all ICD-10-PCS codes and their attributes, as well as functions to search and filter the dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/procedural}
  \item \url{https://andrewallenbruce.github.io/procedural/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/procedural/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
