% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedural-package.R
\docType{package}
\name{procedural-package}
\alias{procedural}
\alias{procedural-package}
\title{procedural: Tidy ICD-10-PCS interface}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andrewallenbruce.github.io/procedural/}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrewbruce.himni@gmail.com} [copyright holder]

}
\keyword{internal}
