% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{sections}
\alias{sections}
\title{ICD-10-PCS Sections}
\usage{
sections(x = NULL)
}
\arguments{
\item{x}{ICD-10-PCS section character, an alphanumeric code of length 1.
If \code{NULL} (default), returns all 17 sections.}
}
\value{
\code{\link[=tibble]{tibble()}}
}
\description{
ICD-10-PCS Sections
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
sections()
sections("0")
sections("2")
sections("X")
\dontshow{\}) # examplesIf}
}
