% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systems.R
\name{systems}
\alias{systems}
\title{ICD-10-PCS Body Systems}
\usage{
systems(x = NULL)
}
\arguments{
\item{x}{ICD-10-PCS body systems character, an alphanumeric code of length 1.
If \code{NULL} (default), returns all 114 systems.}
}
\value{
\code{\link[=tibble]{tibble()}}
}
\description{
ICD-10-PCS Body Systems
}
\examples{
systems()
systems("0")
systems("2")
systems("X")
}
