% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tables}
\alias{tables}
\title{ICD-10-PCS Tables}
\usage{
tables(x = NULL)
}
\arguments{
\item{x}{1 to 3-character string representing an ICD-10-PCS table.
If \code{NULL} (default), returns all 895 tables.}
}
\value{
\code{\link[=tibble]{tibble()}}
}
\description{
ICD-10-PCS Tables
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tables()
tables("0")
tables("00")
tables("00X")
\dontshow{\}) # examplesIf}
}
